Updating Images folder and TreasureHuntWorld code
Updating Images folder. Added all the images for the green-foot folder
that are used in the world.
TreasureHuntWorld code. Creating routes to all the other islands. This
is setting up the route and also assigning the target island that
particular island.
This is the internal logic for the world of the game in greenfoot. I
created objects of other OtherIslands and objects for RouteA and
RouteB. Then coded the logic of sending the RouteA to it’s target
island.